[refactor] #151 mv external related classes to external package & add misc to api
