+++ src/main/java/org/kkumulkkum/server/api/auth/dto/SocialUserDto.java
@@ -1,0 +1,8 @@
+package org.kkumulkkum.server.api.auth.dto;
+
+public record SocialUserDto(String platformId, String email) {
+
+    public static SocialUserDto of(String platformId, String email) {
+        return new SocialUserDto(platformId, email);
+    }
+}
