+++ src/main/java/org/kkumulkkum/server/api/auth/service/AuthService.java
@@ -15,9 +15,7 @@ import org.kkumulkkum.server.domain.userinfo.manager.UserInfoRemover;
 import org.kkumulkkum.server.domain.userinfo.manager.UserInfoRetriever;
 import org.kkumulkkum.server.domain.userinfo.manager.UserInfoSaver;
 import org.kkumulkkum.server.auth.jwt.JwtTokenProvider;
-import org.kkumulkkum.server.auth.openfeign.kakao.dto.SocialUserDto;
-import org.kkumulkkum.server.auth.openfeign.apple.service.AppleService;
-import org.kkumulkkum.server.auth.openfeign.kakao.service.KakaoService;
+import org.kkumulkkum.server.api.auth.dto.SocialUserDto;
 import org.kkumulkkum.server.domain.member.Member;
 import org.kkumulkkum.server.domain.user.Token;
 import org.kkumulkkum.server.domain.user.User;
