+++ src/main/java/org/kkumulkkum/server/api/participant/service/ParticipantService.java
@@ -14,9 +14,9 @@ import org.kkumulkkum.server.api.participant.dto.ParticipantStatusUserInfoDto;
 import org.kkumulkkum.server.api.participant.dto.request.PreparationInfoDto;
 import org.kkumulkkum.server.exception.ParticipantException;
 import org.kkumulkkum.server.exception.code.ParticipantErrorCode;
-import org.kkumulkkum.server.external.FcmService;
-import org.kkumulkkum.server.external.dto.FcmMessageDto;
-import org.kkumulkkum.server.external.enums.FcmContent;
+import org.kkumulkkum.server.external.service.fcm.FcmService;
+import org.kkumulkkum.server.external.service.fcm.dto.FcmMessageDto;
+import org.kkumulkkum.server.external.service.fcm.FcmContent;
 import org.kkumulkkum.server.domain.member.manager.MemberRetreiver;
 import org.kkumulkkum.server.domain.promise.manager.PromiseRemover;
 import org.kkumulkkum.server.domain.promise.manager.PromiseRetriever;
