--- src/main/java/org/kkumulkkum/server/auth/openfeign/apple/dto/ApplePublicKey.java
@@ -1,7 +1,0 @@
-package org.kkumulkkum.server.auth.openfeign.apple.dto;
-
-public record ApplePublicKey(
-        String kty, String kid, String use, String alg, String n, String e, String email
-) {
-
-}
