--- src/main/java/org/kkumulkkum/server/auth/openfeign/apple/dto/ApplePublicKeys.java
@@ -1,16 +1,0 @@
-package org.kkumulkkum.server.auth.openfeign.apple.dto;
-
-import org.kkumulkkum.server.exception.AuthException;
-import org.kkumulkkum.server.exception.code.AuthErrorCode;
-
-import java.util.List;
-
-public record ApplePublicKeys(List<ApplePublicKey> keys) {
-    public ApplePublicKey getMatchesKey(String alg, String kid) {
-        return this.keys
-                .stream()
-                .filter(k -> k.alg().equals(alg) && k.kid().equals(kid))
-                .findFirst()
-                .orElseThrow(() -> new AuthException(AuthErrorCode.INVALID_APPLE_PUBLIC_KEY));
-    }
-}
