--- src/main/java/org/kkumulkkum/server/auth/openfeign/apple/dto/AppleTokenDto.java
@@ -1,14 +1,0 @@
-package org.kkumulkkum.server.auth.openfeign.apple.dto;
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record AppleTokenDto(
-        String accessToken,
-        String tokenType,
-        String expiresIn,
-        String refreshToken,
-        String idToken
-) {
-}
