--- src/main/java/org/kkumulkkum/server/auth/openfeign/kakao/dto/KakaoAccount.java
@@ -1,6 +1,0 @@
-package org.kkumulkkum.server.auth.openfeign.kakao.dto;
-
-public record KakaoAccount(
-        String email
-) {
-}
