--- src/main/java/org/kkumulkkum/server/config/FeignClientConfig.java
@@ -1,10 +1,0 @@
-package org.kkumulkkum.server.config;
-
-import org.kkumulkkum.server.ServerApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@EnableFeignClients(basePackageClasses = ServerApplication.class)
-public class FeignClientConfig {
-}
