--- src/main/java/org/kkumulkkum/server/config/FirebaseConfig.java
@@ -1,35 +1,0 @@
-package org.kkumulkkum.server.config;
-
-import com.google.auth.oauth2.GoogleCredentials;
-import com.google.firebase.FirebaseApp;
-import com.google.firebase.FirebaseOptions;
-import jakarta.annotation.PostConstruct;
-import lombok.extern.slf4j.Slf4j;
-import org.kkumulkkum.server.exception.FirebaseException;
-import org.kkumulkkum.server.exception.code.FirebaseErrorCode;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.io.ClassPathResource;
-
-import java.io.IOException;
-
-@Slf4j
-@Configuration
-public class FirebaseConfig {
-
-    @PostConstruct
-    public void initialize() {
-        try {
-            FirebaseOptions options = FirebaseOptions.builder()
-                    .setCredentials(GoogleCredentials.fromStream(new ClassPathResource("firebase/firebase_service_key.json").getInputStream()))
-                    .build();
-            if (FirebaseApp.getApps().isEmpty()) {
-                FirebaseApp.initializeApp(options);
-                log.info("FirebaseApp initialized {}", FirebaseApp.getInstance().getName());
-            }
-        } catch (IOException e) {
-            log.error("FirebaseApp initialize failed : {}", e.getMessage());
-            throw new FirebaseException(FirebaseErrorCode.NOT_FOUND_FIREBASE_JSON);
-        }
-    }
-}
-
