+++ src/main/java/org/kkumulkkum/server/external/controller/naver/dto/LocationDto.java
@@ -1,0 +1,12 @@
+package org.kkumulkkum.server.external.controller.naver.dto;
+
+public record LocationDto(
+
+        double x,
+        double y,
+        String location,
+        String address,
+        String roadAddress
+) {
+
+}
