+++ src/main/java/org/kkumulkkum/server/external/controller/naver/dto/LocationsDto.java
@@ -1,0 +1,13 @@
+package org.kkumulkkum.server.external.controller.naver.dto;
+
+import java.util.List;
+
+public record LocationsDto (
+        List<LocationDto> locations
+) {
+
+        public static LocationsDto of(List<LocationDto> locations) {
+            return new LocationsDto(locations);
+        }
+
+}
