--- src/main/java/org/kkumulkkum/server/external/dto/LocationDto.java
@@ -1,12 +1,0 @@
-package org.kkumulkkum.server.external.dto;
-
-public record LocationDto(
-
-        double x,
-        double y,
-        String location,
-        String address,
-        String roadAddress
-) {
-
-}
