+++ src/main/java/org/kkumulkkum/server/external/service/kakao/dto/KakaoAccount.java
@@ -1,0 +1,6 @@
+package org.kkumulkkum.server.external.service.kakao.dto;
+
+public record KakaoAccount(
+        String email
+) {
+}
